# Optimize mgmt listener build