--- a/pilot/pkg/networking/core/v1alpha3/networkfilter_test.go
+++ b/pilot/pkg/networking/core/v1alpha3/networkfilter_test.go
@@ -78,7 +78,7 @@ func TestInboundNetworkFilterStatPrefix(t *testing.T) {
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
 
-			env := buildListenerEnv(services)
+			env := buildListenerEnv(services, nil)
 			env.PushContext.InitContext(&env, nil, nil)
 			env.PushContext.Mesh.InboundClusterStatName = tt.statPattern
 
@@ -208,7 +208,7 @@ func TestOutboundNetworkFilterStatPrefix(t *testing.T) {
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
 
-			env := buildListenerEnv(services)
+			env := buildListenerEnv(services, nil)
 			env.PushContext.InitContext(&env, nil, nil)
 			env.PushContext.Mesh.OutboundClusterStatName = tt.statPattern