I also added some basic test for this, as we have none currently

```
name                                 old time/op    new time/op    delta
ListenerGeneration/empty-6             6.98ms ± 3%    5.89ms ± 2%  -15.65%  (p=0.008 n=5+5)
ListenerGeneration/telemetry-6         7.86ms ± 4%    6.71ms ± 3%  -14.58%  (p=0.008 n=5+5)
ListenerGeneration/virtualservice-6    6.70ms ± 1%    5.73ms ± 3%  -14.44%  (p=0.008 n=5+5)

name                                 old alloc/op   new alloc/op   delta
ListenerGeneration/empty-6             2.73MB ± 0%    2.57MB ± 0%   -6.12%  (p=0.008 n=5+5)
ListenerGeneration/telemetry-6         3.13MB ± 0%    2.96MB ± 0%   -5.35%  (p=0.008 n=5+5)
ListenerGeneration/virtualservice-6    2.73MB ± 0%    2.57MB ± 0%   -6.12%  (p=0.008 n=5+5)

name                                 old allocs/op  new allocs/op  delta
ListenerGeneration/empty-6              37.4k ± 0%     30.8k ± 0%  -17.55%  (p=0.000 n=5+4)
ListenerGeneration/telemetry-6          42.4k ± 0%     35.8k ± 0%  -15.49%  (p=0.008 n=5+5)
ListenerGeneration/virtualservice-6     37.4k ± 0%     30.8k ± 0%  -17.55%  (p=0.016 n=4+5)
```